pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "md2docx"
version = "0.1.0"
description = "Convert GitHub-style Markdown to Word using Formato_GIRS template"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "lxml>=5.0.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
md2docx = "md2docx.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
